+++ http/err.go
@@ -1,0 +1,4 @@
+package http
+
+// Err accepts an error which causes the application to stop.
+type Err <-chan error