+++ http/params.go
@@ -1,0 +1,19 @@
+package http
+
+import (
+	"net/http"
+
+	"github.com/go-kit/kit/log"
+	"github.com/goph/fxt"
+	"go.uber.org/dig"
+)
+
+// ServerParams provides a set of dependencies for a tracer constructor.
+type ServerParams struct {
+	dig.In
+
+	Config    *Config `optional:"true"`
+	Handler   http.Handler
+	Logger    log.Logger `optional:"true"`
+	Lifecycle fxt.Lifecycle
+}